# docs/user-openapi.yaml
openapi: 3.0.3
info:
  title: Ecampus Campus Business API
  version: 1.0.0
  description: |
    校园扩展业务用户端 OpenAPI。除文件下载和 WebSocket 外，接口使用统一响应封装。
    新业务 ID 均为十进制字符串；金额单位为分，费率单位为基点；时间为带 +08:00 的 RFC 3339。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/moderation/punishments:
    get:
      tags: [Moderation]
      summary: 我的处罚
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
...
  /api/moderation/appeals:
    get:
      tags: [Moderation]
      summary: 我的申诉
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    SizeQuery:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 15}
...
  responses:
    Success:
      description: 业务成功
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SuccessEnvelope'}
    Error:
      description: 参数或业务失败；语义状态见 body.httpstatus
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, httpstatus, data]
      properties:
        success: {type: boolean, example: true}
        httpstatus: {type: integer, example: 200}
        msg: {type: string, example: 成功}
        data: {}
    ErrorEnvelope:
      type: object
      required: [success, httpstatus, msg]
      properties:
        success: {type: boolean, example: false}
        httpstatus: {type: integer, example: 400}
        msg: {type: string}
        data: {nullable: true}
...